Add basic overlord API files
This commit adds the logging, version and CLI library files, however the
CLI is not yet implemented and displays a warning to this effect.

It does have a full-fledged logging API so that the rest of the library
files can log to both stdout and stderr by using the msg_* and warn_*
functions respectively. The logging level is controlled by the
OVERLORD_LOGLEVEL environment variable and defaults to ERROR (3).