Fix date path computation
On Debian Stretch, date -Isec outputs the timezone with a colon, while
Ubunty Trusty does not output the colon in the timezone. That is, the
output on Debian Stretch is identical to `date +%FT%T%:z`, while on
Ubuntu Trusty, it is `date +%FT%T%z`. Because of this, when creating a
journal entry on Stretch, it mangles the filename, and this causes the
database to go haywire.

Swapping the order of operations avoids this mess and results in the
filename always using YYYY/MM/DD/HHMM.log