Call handler on Group nodes
Prior to this change, Group nodes were expected to not have a handler
function. This change makes it such that if a handler is specified, then
it will be called in the same way that a Leaf node handler is called.
However, the default behavior is preserved, such that if there is no
handler specified, then it will continue to display the help text.