Eliminate NotStarted state from task
The NotStarted state was an initial state, but any state transitions
would never return to this state. It really didn't make sense to keep
this state lying around, so rather than have a dummy NotStarted state,
this commit changes the Paused state to Assigned, and sets the default
state on new tasks to Assigned.