Force due time to be 11:59:59 PM on the due date
This change ensures that any changes to the due date (which is only
specified with day level precision in the UI) get saved as 11:59:59 PM
on that day in the database. This ensures that the due time calculations
don't inadvertently mark a task due on the current date as overdue.